// watcher/src/core/interfaces.ts
export interface GasTxSettings {
  gasLimit: bigint;
  maxFeePerGas: bigint;
  maxPriorityFeePerGas: bigint;
}
// ...
export interface GasAnalysis {
  gasEstimate: bigint;
  gasCostWEI: bigint;
  gasCostUSD: number;
  nativeTokenPriceUSD: number;
  baseFeePerGas: bigint; // latest base fee from latest block
  gasTxSettings: GasTxSettings;
}